Add linux-3.9-at91
Adds recipes for the linux-3.9-at91 branch and changes the machine
configuration to use this as the preferred kernel.  Verified USART0
and USART3 with DMA, USB-B and USB-C, RTC, watchdog and NAND flash.

All the same peripherals are working with the linux-3.6.9-at91
kernel, except the watchdog.  Was able to make the watchdog device
present, but could not force a watchdog reboot.