at91bootstrap: Define AT91BOOTSTRAP_MACHINE
To make it possible to extend this recipe to atmel derivative
machines, lets no use MACHINE variable directly but instead
define AT91BOOTSTRAP_MACHINE which indicates same however
it can be overridden from other layers which makes it possible
to use the recipes for e.g. glomation SBCs which are atmel
based too.

Signed-off-by: Khem Raj <raj.khem@gmail.com>