at91bootstrap: Create common recipe file
Create an include file for at91bootstrap recipe to have a common
configuration for all versions. This way, releasing a new version should
only result in a new recipe with an updated SRC_URI.

Simplify deployment of at91bootstrap images by creating symlinks to the
final image. The deployment directory is also populated with symlinks to
point to the last generated image.

Also, keep only one recipe as recommended by the Yocto project :
https://wiki.yoctoproject.org/wiki/Best_Known_Methods_(BKMs)_for_Package_Updating

Signed-off-by: David Vincent <freesilicon@gmail.com>