sama5d3_xplained: change machine definition due to illegal character
Because of the "_" in the former sama5d3_xplained machine name, we had this
error while building the SDK:
DEBUG: Executing python function do_package_ipk
meta-environment-sama5d3_xplained
*** Error: Package name contains illegal characters, (other than [a-z0-9.+-])

This error is due to presence of uppercase, space or underscore or any other
characters other than [a-z0-9.+-] in MACHINE definition.
So we move to "-" for this machine definition.

Signed-off-by: Nicolas Ferre <nicolas.ferre@atmel.com>