machines: allow qttools-plugins to be empty for building QT5 SDK
Add ALLOW_EMPTY_qttools-plugins = "1" to all MACHINE configuration files.

Without this option, I had the following error when building meta-toolchain-qt5:
* satisfy_dependencies_for: Cannot satisfy the following dependencies for packagegroup-qt5-toolchain-target:
*      qttools-plugins *
* opkg_install_cmd: Cannot install package packagegroup-qt5-toolchain-target.
[..]
Summary: 1 task failed:
  /work/nferre/poky/meta-qt5/recipes-qt/meta/meta-toolchain-qt5.bb, do_populate_sdk

The package qttools-plugins_5.2.1-r0_cortexa5t2hf-vfp-neon.ipk is not created
because it was empty, so force its creation.

This property is related to the machine because the qttools plugins are usually
related to hardware capacities of the platform (GPU).
It won't harm anyway.

Signed-off-by: Nicolas Ferre <nicolas.ferre@atmel.com>