README: re-arrange the directory layout and simplify the build procedure
We don't want that the meta* git trees are embedded in the poky workspace.
So, the build procedure and directory layout (for layers particularly)
is rewritten.
It simplify greatly the build procedure.
Remove also the MACHINE list in the procedure as it is already available
at the beginning of the document.

Signed-off-by: Nicolas Ferre <nicolas.ferre@atmel.com>