qtquick1: need to revert the do_configure_prepend function
To be able to have qtwebkit imports for the qmlbrowser application, we need to
add it to the DEPENDS of qtquick1.

Moreover, the do_configure_prepend() function was also making the imports of
qtwebkit fail, so we removed it.

This recipie depends on include files that are present in meta-qt5.

Signed-off-by: Nicolas Ferre <nicolas.ferre@atmel.com>