QT demo applications: differentiate downloaded archive for each application
If we only use the tag from github without renaming the resulting archive,
we end-up with several v1.x.zip files with different content. This content
obviously doesn't match the md5/sha signature...

Rename the downloaded archive and adapt the source code directory 'S'
accordingly.
Note that we use .tar.gz from github as it's easier to manipulate on a
Linux system.

Signed-off-by: Nicolas Ferre <nicolas.ferre@microchip.com>