sama5d2-xplained: move to Backup+Self-Refresh mode by default
As it is safe to use the B+S-R configuration in AT91Bootstrap for the common
case, we use the B+S-R configuration for the sama5d2-xplained machine by
default.
With newer Linux kernel command line argument, we can configure the kernel
for each sandby/mem PM switches, so we enable the feature in the kernel as
well.

Signed-off-by: Nicolas Ferre <nicolas.ferre@microchip.com>