sama5/defconfig: remove CONFIG_FORCE_MAX_ZONEORDER
Keep the default for the CONFIG_FORCE_MAX_ZONEORDER parameter (to 11 actually)
instead of the 15 value previously.
This was useful for big screen LCD connected on sama5d4 Xplained HDMI
converter. As we now have a separated defconfig for this board, we can
keep the default for sama5 generic defconfig.
This will prevent the boot message on SOM1 EK 128 MiB board:
"cma: Failed to reserve 64 MiB".

Suggested-by: Claudiu Beznea <claudiu.beznea@microchip.com>
Signed-off-by: Nicolas Ferre <nicolas.ferre@microchip.com>