atmel-firmware: remove additional wilc1000 firmware to avoid confusion
Because the Microchip wireless team is maintaining an out of tree driver and
firmware for this module, remove this particular firmware that is hosted on a
removed github tree.
This would avoid confusion.  For any question about this driver, please contact
the Microchip wireless support team or your Microchip representative.

Signed-off-by: Nicolas Ferre <nicolas.ferre@microchip.com>