at91bootstrap: Add dependcy with bc-native
"bc" binary is used in "get_sram_size.sh" script.
With Rocko's branch, it fails to find this binary because now,
sysroots and native-sysroot are available in WORDIR's package. As
the recipe did not have a dependency with "bc", it did not copy
it and failed to find it.

Solved it by adding dependency with bc-native.

Signed-off-by: Mylène Josserand <mylene.josserand@free-electrons.com>