u-boot-atmel: Fix build issue under rocko
In rocko's branch, u-boot fails with an HOSTCC not correctly
configured (set to "cc"). Set it to BUILD_CC variable as it
is done in OpenEmebedded-core's u-boot.inc.
U-boot needs dtc to compile Device Tree blobs. Add a dependency
with "dtc-native" to have this binary available in sysroot-native
folder.
In case of at91sama9x machine, u-boot needs "bc" tool to add the
dependency to bc-native.

Signed-off-by: Mylène Josserand <mylene.josserand@free-electrons.com>