bump applications to new revisions for linux4sam 5.8
This is a big update of all apps for preparing linux4sam 5.8.
Most of the changes are due to the removal of weston replaced by DRM/KMS in QT.
This also changes the way we start applications from the application launcher.

libplanes => 0.0.3
applicationlauncher 1.9 => 1.10
homeautomation 1.9 => 1.10
iocontrol 1.0 => 1.1
minehunt 1.6 => 1.7
qmlbrowser 1.6 => 1.7
samegame 1.5 => 1.6
smartrefrigerator 1.5 => 1.6
videoplayer 1.8 =>  1.9
gstreamer1.0-plugins-hantro 1.2 => 1.3
ptc-examples 1.0 => 1.1

Signed-off-by: Nicolas Ferre <nicolas.ferre@microchip.com>