linux-at91: sam9x60_defconfig: Compile CFG80211 as module
CFG8211 loads the regulatory database as firmware file. Since this is
available only after the rootfs is mounted, we get an error when the
built-in cfg80211 looks for the firmware. Compiling it as a module will
make cfg80211 probe after the rootfs is initialized and the firmware
will be available.

Signed-off-by: Codrin Ciubotariu <codrin.ciubotariu@microchip.com>