kernel defconfig: remove ATMEL_TCLIB as it's not used
Remove this kernel option that is not used anymore for having TCB timer
clocksource clockevent support.
Only use of this "lib" is PWM-TCB which is not used on our boards and needs
source code update. PWM_ATMEL_TCB is thus removed as well.
at91sam9 defconfig is not touched.

Tested on sama5d2 xplained.

Signed-off-by: Nicolas Ferre <nicolas.ferre@microchip.com>