cryptoauthlib: Moved recipe from meta-aws dynamic layer
Currently the cryptoauthlib recipe is only enabled if meta-aws is
included.  There are use cases where crytpoauthlib may be used with PKI
that isn't related to AWS Greengrass.

The recipes-security directory in the aws dynamic layer was moved to the
base directory so cryptoauthlib can be used without the meta-aws layer.

Signed-off-by: Colin McAllister <colinmca242@gmail.com>