python-cryptoauthlib: Removed meta-aws dependency
Currently the python-cryptoauthlib recipe is only enabled if meta-aws is
included. There are use cases where python-cryptoauthlib may be used
with PKI that isn't related to AWS Greengrass.

The python-cryptoauthlib in the aws dynamic layer was moved to the
base directory so python-cryptoauthlib can be used without the meta-aws
layer.

Signed-off-by: Colin McAllister <colinmca242@gmail.com>