cryptoauthlib: Changed eudev requirement to default value
Currently the cryptoauthlib recipe sets eudev as a requirement with
basic variable setting.  As pointed out in issue #172 this breaks
cryptoauthlib's compatability if systemd is included with your build.
This is because eudev is not compatible with systemd.

As pointed out in issue #172, this can be resolved by changing the
requirement to udev.  Instead of changing the requirement to udev,
eudev is set as a default value for requirments.  If a developer
downstream wants to use cryptoauthlib on a build with systemd, they
may create a bbappends file that sets requirements to udev with basic
variable setting.

Signed-off-by: Colin McAllister <colinmca242@gmail.com>