dt-overlay-at91: Implement logic for using kernel+initramfs when selected
Currently, zImage is hardcoded into .its files, this is limiting when
user is in need of using a kernel bundled with initramfs image, this is
nicely reflected with INITRAMFS_IMAGE_BUNDLE from config metadata and
the image name is also consistent which is
KERNEL_IMAGETYPE-initramfs-MACHINE.bin, therefore emit this into the
relevant .its file when initramfs is being used.

This would not come in picuture when INITRAMFS_IMAGE_BUNDLE is not set
hence the default behavior remains as it is.

Signed-off-by: Khem Raj <raj.khem@gmail.com>