dpkg: Remove dpkg DEB_HOST_ARCH append as it is not needed anymore
While this was for sure needed in the past, since quite a long time
this is not needed anymore as it is already part of the standard
poky receipe.
Actually with this patch it is done both on target and native so
it can produce more rebuilds than needed (see more details in
commit f420227ad709b2ab2b5b34f0a2375e7a945640c5 in poky).
So this commit removes it to clean up / simplify the meta-atmel
repository and additionally reduce unneeded rebuilds when using
more MACHINE values (ie. I was building SD/non-SD versions).

Signed-off-by: Federico Pellegrin <fede@evolware.org>