flashing.txt: update and modify meaning of documentation
As we don't provide the outdated flasher.sh script anymore, we repurpose
this documentation to the first step of the flashing process.
I keep it because it's useful information gathered in one place and can
be a good starting point for connecting to comprehensive documentation
either on datasheet or linux4sam.org website.

Signed-off-by: Nicolas Ferre <nicolas.ferre@microchip.com>