linux-at91: configs: Enable crypto software implementations
It's a good idea to enable at least the same amount of algs as
the hardware IPs are supporting. Provide the posibility for the
hw accelerated crypto alg to fallback to software implementation,
in case they need it.
It also helps the crypto test manager to run tests and compare the
results from the software and hardware accelerated implementations.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>