at91bootstrap: add recipe for version 4.0.0
Add recipe for version 4.0.0. This is a new milestone for at91bootstrap.
There are changes that are incompatible with at91bootstrap3, thus creating
variables that have to be set and used in at91bootstrap.inc, to reuse
the deploy and configure tasks.
Deploy is changed because the binaries were moved to build/binaries,
thus created AT91BOOTSTRAP_BIN_PATH to reflect this.
The at91bootstrap.inc will use this variable and each recipe will
define it.
Configure is changed because config files reside now in configs directory.
Created AT91BOOTSTRAP_CONFIG_PATH to reflect this.
Changed bootstrap3 recipe to cope with the variables described above, and
the bootstrap4 recipes will set them as well.
Not all machines from at91bootstrap3 are supported in version 4
Need to pass the sysroot as EXTRA_CC_ARGS because the Makefile in bootstrap
will use quotes to the "$(CC)" variable, which in bootstrap3 includes the
sysroot argument.
AT91bootstrap-3.x will continue to be maintained and updated as the need
arises.
The platforms that support at91bootstrap4 will move to bootstrap4, and older
platforms will continue to stay on bootstrap3.
Bootstrap3 will never be removed as the policy is to still have support for
all the platforms as long as someone is using them.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>
(cherry picked from commit d6e30f2c0e3592ed11f3a4c8380a14d0a9066ba6)
Signed-off-by: Dharma B <dharma.b@microchip.com>