optee-bsp: linux-mchp: add optee support
Enable optee in linux-mchp using a config fragment and
dts patch.

The merge_config.sh available in src tree discards some of the
config fragment's values when allnoconfig option is used, so modify the
option to just merge the config fragment and use make oldconfig to set
the old values.

Signed-off-by: Dharma Balasubiramani <dharma.b@microchip.com>
Reviewed-by: Tudor Ambarus <tudor.ambarus@microchip.com>