egt-benchmark: Use gitsm fetcher
It houses hayai as a submodule, and it fails to build because hayai.h is
missing. The reason is that the sources are not fetched properly.
therefore use gitsm fetcher so it can fetch the hayai submodule and use
it during build

Fixes
'./'`src/main.cpp
src/main.cpp:7:10: fatal error: 'hayai.hpp' file not found
         ^~~~~~~~~~~
1 error generated.
make[1]: *** [Makefile:513: src/benchmark-main.o] Error 1

Signed-off-by: Khem Raj <raj.khem@gmail.com>