meta-mchp-common: recipes-bsp: dt-overlay-mchp: Add support for FIT image
Update the dt-overlay-mchp recipe to compile both DTBOs and ITS files to
generate the final ITB image.

Replace devicetree inheritance with deploy for better control over
deployment.  Override the ITB target in the Makefile to create a FIT image
if an ITS file is present.  This change ensures alignment with the
dt-overlay-mchp repository's intended functionality.

Signed-off-by: Dharma Balasubiramani <dharma.b@microchip.com>
Tested-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Signed-off-by: Hari Prasath Gujulan Elango <hari.prasathge@microchip.com>