meta-mchp-mpu: recipes-graphics: update pixman package
Revert a patch that decreases EGT performances.
The move to pixman 0.40 decreases the score of some EGT benchmarks. This is
something that has been observed with Buildroot as well. This regression
comes from a patch in pixman. It has been reported on the pixman mailing
list, but it didn't ring a bell. So revert this patch.

Signed-off-by: Dharma Balasubiramani <dharma.b@microchip.com>