egt: fix audio files installation
There are two providers for audio files: libegt and egt-media, and it
causes some troubles. So to have only one provider, libegt is
responsible for the audio files as there are already in its repository.
egt-media is the provider of the video files as there are not in the
libegt repository.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>