polarfire-soc: recipes-core: add misc files, init manager configs and images
Add a generic fstab configuration that works with the growfs systemd
service.

Add systemd configurations such as repart config file to extend rootfile
system on first boot.

Extend the core-image-minimal-mtdutils recipe to be used for qspi nor
and flash. A small raw image with .mtd format will be generated for
mpfs-icicle-kit-es-nand and mpfs-icicle-kit-es-nor machines.

Add init-ifupdown configuration for sysvinit (use for icicle-kit-es-nor)

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>