meta-mchp-polarfire-soc: recipes-mchp: add gateware script to packagegroup-mchp-apps
The Icicle Kit with production silicon supports updating the FPGA
fabric using the PolarFire SoC Auto Update feature.

Add the polarfire-soc-linux-examples-gateware to the packagegroup-mchp-
apps to include the gateware example script.

Note: The Icicle Kit with Engineering Sample (ES) silicon (denoted as -es)
does not support writing to flash from the MSS. As a result,
In-Application Programming (IAP) and auto update features are only
supported on production silicon.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Reviewed-by: Eoin Dickson <eoin.dickson@microchip.com>