meta-mchp-polarfire-soc: update sample conf template to use mchp distro
Update the sample local.conf and bblayers to use the Microchip Distro.

Enable the "mchpcore" FEATURE to opt-in to the base configurations
provided by the mchp distro.

Also set the MCHP_RELEASE_VERSION and MCHP_RELEASE_URL to provide
a message of the day (motd) with release information when login in to
the login prompt.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>