meta-mchp-mpu-apps: move EGT apps to separate packagegroup
The packagegroup `packagegroup-mchp-apps-egt` was previously included in
`packagegroup-mchp-apps`, which caused EGT applications to be pulled into
headless images as well. This is not intended, since EGT apps are only
needed for graphics-based images.

Rename it to `packagegroup-mchp-egt-apps` to clearly separate it from the
generic apps group, and updated `mchp-graphics-image.bb` to include the new
packagegroup instead of `mchp-egt-demo-init`.

Also remove the duplicated packagegroup.

Reported-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Signed-off-by: Dharma Balasubiramani <dharma.b@microchip.com>